Refactor API docs and add perms against updating description
