+++ studio/components/interfaces/Docs/RpcContent.tsx
@@ -1,0 +1,97 @@
+import { PermissionAction } from '@supabase/shared-types/out/constants'
+
+import { checkPermissions } from 'hooks'
+import Snippets from 'components/to-be-cleaned/Docs/Snippets'
+import CodeSnippet from 'components/to-be-cleaned/Docs/CodeSnippet'
+import Param from 'components/to-be-cleaned/Docs/Param'
+import Description from 'components/to-be-cleaned/Docs/Description'
+
+/**
+ * TODO: need to support rpc with the same name and different params type
+ */
+const RpcContent = ({
+  autoApiService,
+  rpcId,
+  rpcs,
+  paths,
+  selectedLang,
+  refreshDocs,
+  showApiKey,
+}: any) => {
+  const meta = rpcs[rpcId]
+  const pathKey = `/rpc/${rpcId}`
+  const path = paths && pathKey in paths ? paths[pathKey] : undefined
+  const keyToShow = !!showApiKey ? showApiKey : 'SUPABASE_KEY'
+
+  if (!path) return null
+
+  const {
+    post: { parameters, summary },
+  } = path
+  const rpcParamsObject =
+    parameters && parameters[0] && parameters[0].schema && parameters[0].schema.properties
+      ? parameters[0].schema.properties
+      : {}
+  const rpcParams = Object.entries(rpcParamsObject)
+    .map(([k, v]: any) => ({ name: k, ...v }))
+    .filter((x) => !!x.name)
+  const paramList = rpcParams.map((x) => x.type).join(', ')
+
+  const canUpdateDescription = checkPermissions(PermissionAction.TENANT_SQL_ADMIN_WRITE, 'tables')
+
+  return (
+    <>
+      <h2 className="text-scale-1200 mt-0">
+        <span className="px-6 text-2xl">{meta.id}</span>
+      </h2>
+
+      <div className="doc-section">
+        <article className="text ">
+          <Description
+            content={summary}
+            metadata={{ rpc: `${rpcId} (${paramList})` }}
+            onChange={refreshDocs}
+            disabled={!canUpdateDescription}
+          />
+        </article>
+        <article className="code">
+          <CodeSnippet
+            selectedLang={selectedLang}
+            snippet={Snippets.rpcSingle({
+              rpcName: rpcId,
+              // @ts-ignore
+              rpcCamelCase: meta.camelCase,
+              rpcParams: rpcParams,
+              apiKey: keyToShow,
+              endpoint: autoApiService.endpoint,
+            })}
+          />
+        </article>
+      </div>
+      {rpcParams.length > 0 && (
+        <div>
+          <h3 className="text-scale-1200 mt-0 px-6 capitalize">Function Arguments</h3>
+          {rpcParams.map((x) => {
+            return (
+              <div className="doc-section">
+                <article className="text ">
+                  <Param
+                    key={x.name}
+                    name={x.name}
+                    type={x.type}
+                    format={x.format}
+                    required={true}
+                    description={false}
+                  />
+                </article>
+                <article className="code"></article>
+              </div>
+            )
+          })}
+        </div>
+      )}
+    </>
+  )
+}
+
+export default RpcContent
