+++ studio/components/interfaces/Docs/index.ts
@@ -1,0 +1,5 @@
+import GeneralContent from './GeneralContent'
+import ResourceContent from './ResourceContent'
+import RpcContent from './RPCContent'
+
+export { GeneralContent, ResourceContent, RpcContent }
