+++ studio/components/to-be-cleaned/Docs/Param.js
@@ -8,6 +8,7 @@ const Param = ({
   required,
   description,
   metadata = {},
+  canUpdateDescription = false,
   onDesciptionUpdated = () => {},
 }) => {
   return (
@@ -26,8 +27,12 @@ const Param = ({
       {description !== false && (
         <>
           <label className="uppercase text-xs font-mono text-scale-900">Description</label>
-
-          <Description content={description} metadata={metadata} onChange={onDesciptionUpdated} />
+          <Description
+            content={description}
+            metadata={metadata}
+            onChange={onDesciptionUpdated}
+            disabled={!canUpdateDescription}
+          />
         </>
       )}
     </>
